requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
huggingface_hub>=0.23.0
requests>=2.31.0
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-embeddings-huggingface>=0.3.1
llama-index-readers-web>=0.2.2
llama-index-readers-file>=0.2.2
pypdf>=4.2.0
tiktoken>=0.7.0
nltk>=3.9.1
backoff>=2.2.1